pyproject.toml
[build-system]
requires = ["hatchling>=1.21"]
build-backend = "hatchling.build"

[project]
name = "unslop"
version = "0.1.0"
description = "A simple starter framework for the unslop Python library."
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "unslop" }]
license = { file = "LICENSE" }
keywords = ["template", "library"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
dependencies = [
  "openai>=1.0.0",
  "google-genai>=1.0.0",
  "pathspec>=0.12.1",
  "pywebview>=4.3.2",
  "python-dotenv>=1.0.0",
  "tree-sitter-language-pack>=0.13.0",
  "anthropic>=0.40.0",
]

[project.optional-dependencies]
dev = [
  "black>=24.3.0",
  "ruff>=0.3.0",
  "pytest>=8.1.0",
  "pytest-asyncio>=0.23.0",
]

[project.urls]
Homepage = "https://example.com/unslop"
Issues = "https://example.com/unslop/issues"

[project.scripts]
unslop = "unslop.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/llm", "src/agents", "src/viewer"]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.ruff]
line-length = 88
target-version = "py310"
select = ["E", "F", "B", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"
pythonpath = ["src"]

[tool.pytest_asyncio]
mode = "auto"
